==> 03_Regression_Analyses.Rmd <==
---
title: "Regression Analyses"
author: "Jaewon Royce Choi"
date: "6/19/2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)
```

In this document, I will set up, run, and examine several multiple regression models. In general, the models aim to investigate the relationhips between **broadband** and **entrepreneurship**. Specifically, **how broadband influences entrepreneurship**. Furthermore, the models will explore how different measures of broadband, specifically measures of availability and adoption, explain entrepreneurial activities differently. Finally, I will explore how broadband as well as entrepreneurial factors contribute to the overall prosperity.

```{r Load packages & Get the dataset, results='hide'}

## Load packages
library(tidyverse)
library(ggplot2)
library(gridExtra)
#install.packages("stargazer", dependencies = T)    # A package for better looking tables for results
library(stargazer)

## Read in dataset
tx_bb_entrepreneur_merged_v2 <- read_csv("https://raw.githubusercontent.com/jwroycechoi/broadband-entrepreneurship/master/Datasets/Broadband-Entrepreneurship-TX-merged_v2.csv")

tx_bb_entrepreneur_merged_v2 <- tx_bb_entrepreneur_merged_v2 %>% 
  mutate(chg_pct_50_est_cbp_2012_2018 = pct_50_est_cbp_2018 - pct_50_est_cbp_2012,
         chg_pct_10_est_cbp_2012_2018 = pct_10_est_cbp_2018 - pct_10_est_cbp_2012)

```

## Broadband and Entrepreneurship

First, let's take another look at some of the key variables.

> Entrepreneurship

* `IRR2010`: Rural index (larger = rural)
* `pct_change_pro_emp_2010_2017`: Change in % of sole proprietors b/w 2010-2017
* `chg_pct_10_est_cbp_2012_2018`: Change in % of establishments with less than 10 employees b/w 2012-2017
* `chg_pct_nonfarmneest_nemp_2012_2018`: Change in % of nonfarm nonemployer b/w 2012-2018
* `pct_chg_bea_2012_2018`: Change in % of nonfarm proprietors b/w 2012-2018
* `venturedensity_mean`: Venture density
* `highlyactivevd_mean`: Highly active venture density

> Broadband

* `pct_broadband_FCC`: FCC broadband availability in % (2017)
* `pct_broadband_MS`: MS broadband availability in % (2019)
* `pct_fixed_acs_2018`: ACS broadband subscription in $ (2018)
* `pct_broadband_mlab`: M-Lab broadband measure (2019)

### Broadband's Effect on Entrepreneurial Outcome

Here I investigate **models that estimate broadband variables' effect on the entrepreneurial outcome.** Several different broadband measures are in the dataset. I will first model **each measure's relationship with the entrepreneurial outcome separately.**

Entrepreneurial outcomes can be represented by several variables in this dataset. As the ASU research team's white paper argued, the venture density provided by GoDaddy activities reflects actual entrepreneurial activities in the form of online domain creation. In the dataset, we also have Texas state's sole proprietor statistics. These variables reflect small and medium sized businesses in each county.

#### Control Variables

For control variables, various demographic measures derived from the ACS estimates are available in the dataset. Furthermore, the ruralithy of the counties are reflected in the IRR index. Number of population, and regional intustry characteristics will be also included as control variables.

## FCC Broadband Availability

```{r FCC Regression Analyses, results='asis'}
#### Model 1 ####

## DV: Venture Density, Highly Active VD, Change in Proprietors Share
## IV: Broadband Availability (FCC)

model_fcc_vd <- lm(venturedensity_mean ~ pct_broadband_FCC + IRR2010 + pctagriculture + pctconstruction + pctwholesale + pctretail + pcttransportation + pctinformation_tech + pctfinance + pctprofessional + pcteducation + pctother_occupation + pctpublic + pcthighschool + pctcollege + pctmillennial + pctgenx + pctbabyboomer,
              data = tx_bb_entrepreneur_merged_v2)

model_fcc_hvd <- lm(highlyactive_vd_mean ~ pct_broadband_FCC + IRR2010 + pctagriculture + pctconstruction + pctwholesale + pctretail + pcttransportation + pctinformation_tech + pctfinance + pctprofessional + pcteducation + pctother_occupation + pctpublic + pcthighschool + pctcollege + pctmillennial + pctgenx + pctbabyboomer,
              data = tx_bb_entrepreneur_merged_v2)

model_fcc_pro <- lm(pct_chg_bea_2012_2018 ~ pct_broadband_FCC + IRR2010 + pctagriculture + pctconstruction + pctwholesale + pctretail + pcttransportation + pctinformation_tech + pctfinance + pctprofessional + pcteducation + pctother_occupation + pctpublic + pcthighschool + pctcollege + pctmillennial + pctgenx + pctbabyboomer,
              data = tx_bb_entrepreneur_merged_v2)
## Generate Table
stargazer(model_fcc_vd, model_fcc_hvd, model_fcc_pro, title = "Regression Results", type = "html",
          align = TRUE, dep.var.labels = c("Venture Density","Highly Active VD", "Proprietors Share"),
          covariate.labels = c("FCC Broadband Availability","Rurality Index (2010)","Population","Black","Native American","Asian","Hispanic","Agriculture","Construction","Wholesale","Retail","Transportation","IT","Finance","Professional","Education","Other occupation","Public","Highschool","College","Millenials","GenX","Babyboomer","Foreign Born","Median Income (2017)"),
          omit.stat = c("ser"), no.space = TRUE)

```

## MS Broadband Availability

```{r MS Regression Analyses, results='asis'}
#### Model 2 ####

## DV: Venture Density, Highly Active VD, Change in Proprietors Share
## IV: Broadband Availability (MS)

model_ms_vd <- lm(venturedensity_mean ~ pct_broadband_MS + IRR2010 + pctagriculture + pctconstruction + pctwholesale + pctretail + pcttransportation + pctinformation_tech + pctfinance + pctprofessional + pcteducation + pctother_occupation + pctpublic + pcthighschool + pctcollege + pctmillennial + pctgenx + pctbabyboomer,
              data = tx_bb_entrepreneur_merged_v2)

model_ms_hvd <- lm(highlyactive_vd_mean ~ pct_broadband_MS + IRR2010 + pctagriculture + pctconstruction + pctwholesale + pctretail + pcttransportation + pctinformation_tech + pctfinance + pctprofessional + pcteducation + pctother_occupation + pctpublic + pcthighschool + pctcollege + pctmillennial + pctgenx + pctbabyboomer,
              data = tx_bb_entrepreneur_merged_v2)

model_ms_pro <- lm(pct_chg_bea_2012_2018 ~ pct_broadband_MS + IRR2010 + pctagriculture + pctconstruction + pctwholesale + pctretail + pcttransportation + pctinformation_tech + pctfinance + pctprofessional + pcteducation + pctother_occupation + pctpublic + pcthighschool + pctcollege + pctmillennial + pctgenx + pctbabyboomer,
              data = tx_bb_entrepreneur_merged_v2)
## Generate Table
stargazer(model_ms_vd, model_ms_hvd, model_ms_pro, title = "Regression Results", type = "html",
          align = TRUE, dep.var.labels = c("Venture Density","Highly Active VD", "Proprietors Share"),
          covariate.labels = c("MS Broadband Availability","Rurality Index (2010)","Population","Black","Native American","Asian","Hispanic","Agriculture","Construction","Wholesale","Retail","Transportation","IT","Finance","Professional","Education","Other occupation","Public","Highschool","College","Millenials","GenX","Babyboomer","Foreign Born","Median Income (2017)"),
          omit.stat = c("ser"), no.space = TRUE)
```

## ACS Broadband Subscription Level

```{r ACS Regression Analyses, results='asis'}
#### Model 3 ####

## DV: Venture Density, Highly Active VD, Change in Proprietors Share
## IV: Broadband Subscription (ACS)

model_acs_vd <- lm(venturedensity_mean ~ pct_fixed_acs_2018 + IRR2010 + pctagriculture + pctconstruction + pctwholesale + pctretail + pcttransportation + pctinformation_tech + pctfinance + pctprofessional + pcteducation + pctother_occupation + pctpublic + pcthighschool + pctcollege + pctmillennial + pctgenx + pctbabyboomer,
              data = tx_bb_entrepreneur_merged_v2)

model_acs_hvd <- lm(highlyactive_vd_mean ~ pct_fixed_acs_2018 + IRR2010 + pctagriculture + pctconstruction + pctwholesale + pctretail + pcttransportation + pctinformation_tech + pctfinance + pctprofessional + pcteducation + pctother_occupation + pctpublic + pcthighschool + pctcollege + pctmillennial + pctgenx + pctbabyboomer,
              data = tx_bb_entrepreneur_merged_v2)

model_acs_pro <- lm(pct_chg_bea_2012_2018 ~ pct_fixed_acs_2018 + IRR2010 + pctagriculture + pctconstruction + pctwholesale + pctretail + pcttransportation + pctinformation_tech + pctfinance + pctprofessional + pcteducation + pctother_occupation + pctpublic + pcthighschool + pctcollege + pctmillennial + pctgenx + pctbabyboomer,
              data = tx_bb_entrepreneur_merged_v2)
## Generate Table
stargazer(model_acs_vd, model_acs_hvd, model_acs_pro, title = "Regression Results", type = "html",
          align = TRUE, dep.var.labels = c("Venture Density","Highly Active VD", "Proprietors Share"),
          covariate.labels = c("ACS Broadband Subscription","Rurality Index (2010)","Population","Black","Native American","Asian","Hispanic","Agriculture","Construction","Wholesale","Retail","Transportation","IT","Finance","Professional","Education","Other occupation","Public","Highschool","College","Millenials","GenX","Babyboomer","Foreign Born","Median Income (2017)"),
          omit.stat = c("ser"), no.space = TRUE)

```

## M-Lab Broadband Availability

```{r M-Lab Regression Analyses, results='asis'}
#### Model 4 ####

## DV: Venture Density, Highly Active VD, Change in Proprietors Share
## IV: Broadband Availability (M-Lab)

model_mlab_vd <- lm(venturedensity_mean ~ pct_broadband_mlab + IRR2010 + pctagriculture + pctconstruction + pctwholesale + pctretail + pcttransportation + pctinformation_tech + pctfinance + pctprofessional + pcteducation + pctother_occupation + pctpublic + pcthighschool + pctcollege + pctmillennial + pctgenx + pctbabyboomer,
              data = tx_bb_entrepreneur_merged_v2)

model_mlab_hvd <- lm(highlyactive_vd_mean ~ pct_broadband_mlab + IRR2010 + pctagriculture + pctconstruction + pctwholesale + pctretail + pcttransportation + pctinformation_tech + pctfinance + pctprofessional + pcteducation + pctother_occupation + pctpublic + pcthighschool + pctcollege + pctmillennial + pctgenx + pctbabyboomer,
              data = tx_bb_entrepreneur_merged_v2)

model_mlab_pro <- lm(pct_chg_bea_2012_2018 ~ pct_broadband_mlab + IRR2010 + pctagriculture + pctconstruction + pctwholesale + pctretail + pcttransportation + pctinformation_tech + pctfinance + pctprofessional + pcteducation + pctother_occupation + pctpublic + pcthighschool + pctcollege + pctmillennial + pctgenx + pctbabyboomer,
              data = tx_bb_entrepreneur_merged_v2)
## Generate Table
stargazer(model_mlab_vd, model_mlab_hvd, model_mlab_pro, title = "Regression Results", type = "html",
          align = TRUE, dep.var.labels = c("Venture Density","Highly Active VD", "Proprietors Share"),
          covariate.labels = c("M-Lab Broadband","Rurality Index (2010)","Population","Black","Native American","Asian","Hispanic","Agriculture","Construction","Wholesale","Retail","Transportation","IT","Finance","Professional","Education","Other occupation","Public","Highschool","College","Millenials","GenX","Babyboomer","Foreign Born","Median Income (2017)"),
          omit.stat = c("ser"), no.space = TRUE)

```

### Models with Fewer Control Variables

Below, simpler models with fewer, yet important control variables will be explored. In a previous meeting with Dr.Brian Whitacre, he suggested age and education might be meaningful control variables. Here I will use rurality index, population, education, and age cohort variables as control variables.

## FCC Broadband Availability

```{r FCC Simpler Regression Models, results='asis'}
## Fewer Control Variables: Rurality, population, education level, age cohorts
## FCC Broadband
# VD model
model_fcc_vd2 <- lm(venturedensity_mean ~ pct_broadband_FCC + IRR2010 + population + pcthighschool + pctcollege + pctmillennial + pctgenx + pctbabyboomer,
                    data = tx_bb_entrepreneur_merged_v2)

# HVD model
model_fcc_hvd2 <- lm(highlyactive_vd_mean ~ pct_broadband_FCC + IRR2010 + population + pcthighschool + pctcollege + pctmillennial + pctgenx + pctbabyboomer,
                    data = tx_bb_entrepreneur_merged_v2)

# Proprietor model
model_fcc_pro2 <- lm(pct_chg_bea_2012_2018 ~ pct_broadband_FCC + IRR2010 + population + pcthighschool + pctcollege + pctmillennial + pctgenx + pctbabyboomer,
                     data = tx_bb_entrepreneur_merged_v2)

## Generate Table
stargazer(model_fcc_vd2, model_fcc_hvd2, model_fcc_pro2, type = "html", align = T,
          dep.var.labels = c("Venture Density","Highly Active VD", "Proprietors Share"),
          covariate.labels = c("FCC Broadband Availability",
                               "Rurality Index (2010)",
                               "Population","Highschool","College",
                               "Millenials","GenX","Babyboomer"),
          omit.stat = c("ser"), no.space = T)


```

## MS Broadband Availability

```{r MS Simpler Models, results='asis'}
## MS Broadband
# VD model
model_ms_vd2 <- lm(venturedensity_mean ~ pct_broadband_MS + IRR2010 + population + pcthighschool + pctcollege + pctmillennial + pctgenx + pctbabyboomer,
                   data = tx_bb_entrepreneur_merged_v2)

# HVD model
model_ms_hvd2 <- lm(highlyactive_vd_mean ~ pct_broadband_MS + IRR2010 + population + pcthighschool + pctcollege + pctmillennial + pctgenx + pctbabyboomer,
                   data = tx_bb_entrepreneur_merged_v2)

# Proprietor model
model_ms_pro2 <- lm(pct_chg_bea_2012_2018 ~ pct_broadband_MS + IRR2010 + population + pcthighschool + pctcollege + pctmillennial + pctgenx + pctbabyboomer,
                   data = tx_bb_entrepreneur_merged_v2)

## Generate Table
stargazer(model_ms_vd2, model_ms_hvd2, model_ms_pro2, type = "html", align = T,
          dep.var.labels = c("Venture Density","Highly Active VD", "Proprietors Share"),
          covariate.labels = c("MS Broadband Availability",
                               "Rurality Index (2010)",
                               "Population","Highschool","College",
                               "Millenials","GenX","Babyboomer"),
          omit.stat = c("ser"), no.space = T)

```

## ACS Broadband Subscription Level

```{r ACS Simpler Models, results='asis'}
## ACS Broadband
# VD model
model_acs_vd2 <- lm(venturedensity_mean ~ pct_fixed_acs_2018 + IRR2010 + population + pcthighschool + pctcollege + pctmillennial + pctgenx + pctbabyboomer,
                   data = tx_bb_entrepreneur_merged_v2)

# HVD model
model_acs_hvd2 <- lm(highlyactive_vd_mean ~ pct_fixed_acs_2018 + IRR2010 + population + pcthighschool + pctcollege + pctmillennial + pctgenx + pctbabyboomer,
                    data = tx_bb_entrepreneur_merged_v2)

# Proprietor model
model_acs_pro2 <- lm(pct_chg_bea_2012_2018 ~ pct_fixed_acs_2018 + IRR2010 + population + pcthighschool + pctcollege + pctmillennial + pctgenx + pctbabyboomer,
                    data = tx_bb_entrepreneur_merged_v2)

## Generate Table
stargazer(model_acs_vd2, model_acs_hvd2, model_acs_pro2, type = "html", align = T,
          dep.var.labels = c("Venture Density","Highly Active VD", "Proprietors Share"),
          covariate.labels = c("ACS Broadband Subscription",
                               "Rurality Index (2010)",
                               "Population","Highschool","College",
                               "Millenials","GenX","Babyboomer"),
          omit.stat = c("ser"), no.space = T)


```

## M-Lab Broadband Availability

```{r M-Lab Simpler Models, results='asis'}
## ACS Broadband
# VD model
model_mlab_vd2 <- lm(venturedensity_mean ~ pct_broadband_mlab + IRR2010 + population + pcthighschool + pctcollege + pctmillennial + pctgenx + pctbabyboomer,
                   data = tx_bb_entrepreneur_merged_v2)

# HVD model
model_mlab_hvd2 <- lm(highlyactive_vd_mean ~ pct_broadband_mlab + IRR2010 + population + pcthighschool + pctcollege + pctmillennial + pctgenx + pctbabyboomer,
                    data = tx_bb_entrepreneur_merged_v2)

# Proprietor model
model_mlab_pro2 <- lm(pct_chg_bea_2012_2018 ~ pct_broadband_mlab + IRR2010 + population + pcthighschool + pctcollege + pctmillennial + pctgenx + pctbabyboomer,
                    data = tx_bb_entrepreneur_merged_v2)

## Generate Table
stargazer(model_mlab_vd2, model_mlab_hvd2, model_mlab_pro2, type = "html", align = T,
          dep.var.labels = c("Venture Density","Highly Active VD", "Proprietors Share"),
          covariate.labels = c("M-Lab Broadband",
                               "Rurality Index (2010)",
                               "Population","Highschool","College",
                               "Millenials","GenX","Babyboomer"),
          omit.stat = c("ser"), no.space = T)

```
