==> 06_Entrepreneurship_Measurements.Rmd <==
---
title: "Entrepreneurship Measurements"
author: "Jaewon Royce Choi"
date: "7/6/2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, fig.align = 'center', warning = FALSE)
```

Here I will explore different potential entrepreneurship measurements available. In the current dataset, we have several different potential measures of entrepreneurship:

* **Sole Proprietors Share**: % of sole proprietors in total employment _(2000, 2010, 2017)_
* **Change in Sole Proprietors Share**: Change in the % of sole proprietors in total employment _(b/w 2000, 2010, 2017 each)_
* **Venture Density**: Online venture activities based on GoDaddy database (average of data points available throughout _2018 and 2019_)
* **Highly Active Venture Density**: Highly active cluster of venture activities based on GoDaddy database (average of data points available throughout _2018 and 2019_)
* **Percent of County Business Establishment with <=50 Employees**: % of business establishments with less than 50 employees _(County Business Dynamics dataset from 2012,2017,2018)_
* **Percent of County Business Establishment with <=10 Employees**: % of business establishment with less than 10 employees _(County Business Dynamics dataset from 2012,2017,2018)_
* **Percent of Nonfarm Nonemployer Establishments**: % of nonfarm nonemployer establishment _(Nonemployer Statistics dataset from 2012,2017,2018)_
* **Percent of Nonfarm Proprietors**: % of nonfarm proprietors _(Bureau of Economic Analysis dataset from 2012,2018)_
* **Firms with Paid Employees**: Number of firms with paid employees _(Annual Business Survey dataset from 2012,2017)_
* **Change in Average Nonemployer Establishment**: Change of the average number of nonemployer establishments from 2017 to 2018

Throughout this page, I will explore descriptive statistics of these measures, distributions of the measures, correlations between these measures, and also how these measures look like on the map geographically.

## Descriptive Statistics of Broadband Measures

```{r Impoart Libraries, results='hide'}
library(leaflet)
library(tidyverse)
library(RColorBrewer)
library(tigris)
library(ggplot2)
library(sp)
library(stargazer)
library(gridExtra)

#### Import the merged dataset V2 ####
tx_bb_entrepreneur_merged_v2 <- read_csv("https://raw.githubusercontent.com/jwroycechoi/broadband-entrepreneurship/master/Datasets/Broadband-Entrepreneurship-TX-merged_v2.csv")

tx_bb_entrepreneur_merged_v2 <- tx_bb_entrepreneur_merged_v2 %>% 
  mutate(chg_pct_50_est_cbp_2012_2018 = pct_50_est_cbp_2018 - pct_50_est_cbp_2012,
         chg_pct_10_est_cbp_2012_2018 = pct_10_est_cbp_2018 - pct_10_est_cbp_2012)

## Set up spatial dataframe
tx_county <- counties("Texas")
tx_bb_entrepreneur_merged_v3 <- sp::merge(tx_county, tx_bb_entrepreneur_merged_v2,
                                          by.x = "GEOID", by.y = "FIPS")
```

```{r Descriptive Statistics, results='asis'}
#### Descriptive Statistics ####
tx_bb_entrepreneur_merged_v2 %>% 
  select(pct_proprietors_employment_2017, pct_50_est_cbp_2018, pct_10_est_cbp_2018, pct_nonfarmneest_nemp_2018, pct_nonfarm_bea_2018, venturedensity_mean, highlyactive_vd_mean) %>% 
  stargazer(as.data.frame(.), type = "html")
```

## Distributions of Broadband Measures

```{r Distributions of Broadband Measures, echo=FALSE}
grid.arrange(
  ggplot(tx_bb_entrepreneur_merged_v2, aes(x = pct_proprietors_employment_2017)) + geom_histogram() + theme_minimal() + xlab("% of Sole Proprietors (2017)"),
  ggplot(tx_bb_entrepreneur_merged_v2, aes(x = pct_50_est_cbp_2018)) + geom_histogram() + theme_minimal() + xlab("% of Establishment\n<50 Employees (2018)"),
  ggplot(tx_bb_entrepreneur_merged_v2, aes(x = pct_10_est_cbp_2018)) + geom_histogram() + theme_minimal() + xlab("% of Establishment\n<10 Employees (2018)"),
  ggplot(tx_bb_entrepreneur_merged_v2, aes(x = pct_nonfarmneest_nemp_2018)) + geom_histogram() + theme_minimal() + xlab("% of Nonfarm\nNonemployer (2018)"),
  ggplot(tx_bb_entrepreneur_merged_v2, aes(x = pct_nonfarm_bea_2018)) + geom_histogram() + theme_minimal() + xlab("% of Nonfarm\nProprietors (2018)"),
  nrow = 2, ncol = 3, top = "Entrepreneurship Measure Distribution (1)"
)

grid.arrange(
  ggplot(tx_bb_entrepreneur_merged_v2, aes(x = pct_change_pro_emp_2010_2017)) + geom_histogram() + theme_minimal() + xlab("Change in % of\nSole Proprietors (2010-2017)"),
  ggplot(tx_bb_entrepreneur_merged_v2, aes(x = chg_pct_10_est_cbp_2012_2018)) + geom_histogram() + theme_minimal() + xlab("Change in % of\nEstablishments (<10 emp) (2012-2018)"),
  ggplot(tx_bb_entrepreneur_merged_v2, aes(x = chg_pct_nonfarmneest_nemp_2012_2018)) + geom_histogram() + theme_minimal() + xlab("Change in % of\nNonfarm Nonemployers (2012-2018)"),
  ggplot(tx_bb_entrepreneur_merged_v2, aes(x = pct_chg_bea_2012_2018)) + geom_histogram() + theme_minimal() + xlab("Change in % of\n Nonfarm Proprietors (2012-2018)"),
  ggplot(tx_bb_entrepreneur_merged_v2, aes(x = chg_firm_2012_2017)) + geom_histogram() + theme_minimal() + xlab("Change in # of\n Firms (2012-2017)"),
  ggplot(tx_bb_entrepreneur_merged_v2, aes(x = venturedensity_mean)) + geom_histogram() + theme_minimal() + xlab("Venture Density"),
  ggplot(tx_bb_entrepreneur_merged_v2, aes(x = highlyactive_vd_mean)) + geom_histogram() + theme_minimal() + xlab("Highly Active\nVenture Density"),
  nrow = 2, ncol = 4, top = "Entrepreneurship Measure Distribution (2)"
)
```


## Correlations b/w Entrepreneurship Measures

```{r Correlation and Relationship Altogether, echo=FALSE}
tx_bb_entrepreneur_merged_v2 %>% 
  select(pct_change_pro_emp_2010_2017, chg_pct_10_est_cbp_2012_2018, chg_pct_nonfarmneest_nemp_2012_2018, pct_chg_bea_2012_2018, chg_firm_2012_2017, venturedensity_mean, highlyactive_vd_mean) %>% 
  PerformanceAnalytics::chart.Correlation(histogram = T)

tx_bb_entrepreneur_merged_v2 %>% 
  select(pct_proprietors_employment_2017, pct_50_est_cbp_2018, pct_10_est_cbp_2018, pct_nonfarmneest_nemp_2018, pct_nonfarm_bea_2018, venturedensity_mean, highlyactive_vd_mean) %>% 
  PerformanceAnalytics::chart.Correlation(histogram = T)
```


## Mapping Different Entrepreneurship Measures

```{r Map Setup, echo=FALSE, results='hide'}
## Create continuous color palette function based on FCC broadband range (0~1) ##
pal <- colorNumeric(palette = "YlOrRd", domain = tx_bb_entrepreneur_merged_v2$pct_change_pro_emp_2010_2017)
pal2 <- colorNumeric(palette = "YlOrRd", domain = tx_bb_entrepreneur_merged_v2$venturedensity_mean)
pal3 <- colorNumeric(palette = "YlOrRd", domain = tx_bb_entrepreneur_merged_v2$highlyactive_vd_mean)
pal4 <- colorNumeric(palette = "YlOrRd", domain = tx_bb_entrepreneur_merged_v2$chg_pct_10_est_cbp_2012_2018)
pal5 <- colorNumeric(palette = "YlOrRd", domain = tx_bb_entrepreneur_merged_v2$chg_pct_nonfarmneest_nemp_2012_2018)
pal6 <- colorNumeric(palette = "YlOrRd", domain = tx_bb_entrepreneur_merged_v2$pct_chg_bea_2012_2018)
pal7 <- colorNumeric(palette = "YlOrRd", domain = tx_bb_entrepreneur_merged_v2$chg_firm_2012_2017)
## Set a popup template ##
popup_tx <- paste0("<strong>", tx_bb_entrepreneur_merged_v3$county,
                   "</strong><br />% Sole Proprietor: ", round(tx_bb_entrepreneur_merged_v3$pct_proprietors_employment_2017*100, digits = 1), "%",
                   "<br />Venture Density: ", round(tx_bb_entrepreneur_merged_v3$venturedensity_mean, digits = 1),
                   "<br />Highly Active VD: ", round(tx_bb_entrepreneur_merged_v3$highlyactive_vd_mean, digits = 1),
                   "<br />Total Employment (2018): ", tx_bb_entrepreneur_merged_v3$emp_cbp_2018,
                   "<br />Total # Firms (2017): ", tx_bb_entrepreneur_merged_v3$firm_2017)
head(popup_tx)
## Create a map object ##
map <- leaflet(tx_bb_entrepreneur_merged_v3)
```

### Change in the Share of Sole Proprietors (2010-2017)

```{r Sole Proprietorship, echo=FALSE}
## Sole Proprietors Share Map ##
map %>% addTiles() %>% setView(lng = -100.000, lat = 31.000, zoom = 6) %>% 
  addPolygons(stroke = F, smoothFactor = 0.2, fillOpacity = 0.9,
                                   color = ~pal(pct_change_pro_emp_2010_2017),
              popup = ~popup_tx) %>% 
  addLegend("bottomright", pal = pal, values = ~pct_change_pro_emp_2010_2017,
            title = "Change in<br>Sole Proprietors Share (%)",
            labFormat = labelFormat(suffix = "%", transform = function(x) 100*x), na.label = "N/A", opacity = 1)
```

### Venture Density in Texas

```{r Venture Density, echo=FALSE}
## Venture Density Map ##
map %>% addTiles() %>% setView(lng = -100.000, lat = 31.000, zoom = 6) %>% 
  addPolygons(stroke = F, smoothFactor = 0.2, fillOpacity = 0.9,
              color = ~pal2(venturedensity_mean),
              popup = ~popup_tx) %>% 
  addLegend("bottomright", pal = pal2, values = ~venturedensity_mean,
            title = "Venture Density", opacity = 1)
```

### Highly Active Venture Density in Texas

```{r Highly Active Venture Density, echo=FALSE}
## Highly Active VD Map ##
map %>% addTiles() %>% setView(lng = -100.000, lat = 31.000, zoom = 6) %>% 
  addPolygons(stroke = F, smoothFactor = 0.2, fillOpacity = 0.9,
              color = ~pal3(highlyactive_vd_mean),
              popup = ~popup_tx) %>% 
  addLegend("bottomright", pal = pal3, values = ~highlyactive_vd_mean,
            title = "Highly Active VD",
            opacity = 1)
```

### Change in Establishments with less than 50 Employees (2012-2018)

```{r Nonemployer Establishment, echo=FALSE}
## Nonemployer Establishment Map ##
map %>% addTiles() %>% setView(lng = -100.000, lat = 31.000, zoom = 6) %>% 
  addPolygons(stroke = F, smoothFactor = 0.2, fillOpacity = 0.9,
              color = ~pal4(chg_pct_10_est_cbp_2012_2018),
              popup = ~popup_tx) %>% 
  addLegend("bottomright", pal = pal4, values = ~chg_pct_10_est_cbp_2012_2018,
            title = "Change in <br><10 Establishments (%)",
            labFormat = labelFormat(suffix = "%", transform = function(x) 100*x), na.label = "N/A",
            opacity = 1)
```

### Change in Nonfarm Nonemployer Establishments (2012-2018)

```{r Nonemployer Establishment, echo=FALSE}
## Nonemployer Establishment Map ##
map %>% addTiles() %>% setView(lng = -100.000, lat = 31.000, zoom = 6) %>% 
  addPolygons(stroke = F, smoothFactor = 0.2, fillOpacity = 0.9,
              color = ~pal5(chg_pct_nonfarmneest_nemp_2012_2018),
              popup = ~popup_tx) %>% 
  addLegend("bottomright", pal = pal5, values = ~chg_pct_nonfarmneest_nemp_2012_2018,
            title = "Change in <br>Nonfarm Nonemployer (%)",
            labFormat = labelFormat(suffix = "%", transform = function(x) 100*x), na.label = "N/A",
            opacity = 1)
```

### Change in Nonfarm Proprietors (2012-2018)

```{r Nonemployer Establishment, echo=FALSE}
## Nonemployer Establishment Map ##
map %>% addTiles() %>% setView(lng = -100.000, lat = 31.000, zoom = 6) %>% 
  addPolygons(stroke = F, smoothFactor = 0.2, fillOpacity = 0.9,
              color = ~pal6(pct_chg_bea_2012_2018),
              popup = ~popup_tx) %>% 
  addLegend("bottomright", pal = pal6, values = ~pct_chg_bea_2012_2018,
            title = "Change in <br>Nonfarm Proprietors (%)",
            labFormat = labelFormat(suffix = "%", transform = function(x) 100*x), na.label = "N/A",
            opacity = 1)
```

### Change in Number of Firms with Paid Employees (2012-2017)

```{r Nonemployer Establishment, echo=FALSE}
## Nonemployer Establishment Map ##
map %>% addTiles() %>% setView(lng = -100.000, lat = 31.000, zoom = 6) %>% 
  addPolygons(stroke = F, smoothFactor = 0.2, fillOpacity = 0.9,
              color = ~pal7(chg_firm_2012_2017),
              popup = ~popup_tx) %>% 
  addLegend("bottomright", pal = pal7, values = ~chg_firm_2012_2017,
            title = "Change in <br># of Firms",
            na.label = "N/A",
            opacity = 1)
```
